review(#1271): batch 2 — promote LoRA ParameterCount math to long
The ParameterCount overrides on 9 LoRA adapters / layers were summing
through int arithmetic before widening to the long return type, so
sufficiently large bases / ranks would wrap silently before the
ParameterCountHelper.ToFlatVectorSize guard saw them. ChainLoRAAdapter
also cached the result in an int field. Promoted everything to long
end-to-end:

- LoRALayer.cs: cast Rows/Columns to long BEFORE the multiplication
  (closes #1271.7Bnv)
- ChainLoRAAdapter: _currentParameterCount is now long; the on-the-fly
  and cached-update paths both accumulate as long; Vector<T>
  allocation goes through ParameterCountHelper.ToFlatVectorSize so a
  >int.MaxValue chain fails fast with the actionable error message
  (closes #1271.7BnU)
- DeltaLoRAAdapter (closes #1271.7BnX)
- GLoRAAdapter (closes #1271.7Bna)
- LoRETTAAdapter — TT-core product accumulator promoted (closes
  #1271.7Bnd)
- NOLAAdapter (closes #1271.7Bni)
- RoSAAdapter — sparseCount cast promoted (closes #1271.7Bnj)
- TiedLoRAAdapter (closes #1271.7Bnn)
- XLoRAAdapter — expert/gating accumulators promoted (closes #1271.7Bnq)

Build clean on net10.0. No behaviour change for in-range models;
the difference only appears once a sum crosses int.MaxValue (where
the old int math silently wrapped, now the long-typed sum reaches
the helper which surfaces the actionable error).

+++ src/LoRA/Adapters/DeltaLoRAAdapter.cs
@@ -131,8 +131,11 @@ public class DeltaLoRAAdapter<T> : LoRAAdapterBase<T>
     {
         get
         {
-            int baseCount = (int)base.ParameterCount; // Base layer + LoRA layer
-            int deltaCount = _deltaWeights.Rows * _deltaWeights.Columns;
+            // long throughout — base layer + LoRA layer + delta weights
+            // can sum past int.MaxValue on foundation-model scales.
+            // Closes #1271.7BnX.
+            long baseCount = base.ParameterCount;
+            long deltaCount = (long)_deltaWeights.Rows * _deltaWeights.Columns;
             return baseCount + deltaCount;
         }
     }
